Created dataframe to extract and store user data and information
Create and execution of a for loop which adds users to the list. This If statement determines if a user has no followers, it skips over to the next. Execute a loop to iterate through the following users to the account. If statement to ensure the list does not contain duplicate values and adds those to the list . Add this user's data to the aforementioned dataframe usersDB.